requirements.txt
future
flake8
mock
pytest
pyyaml==3.11
termcolor==1.1.0
tox
